Replies: Fix that the webhook message won't edit webhook message reply by mistake
As webhook message replies are added in two lists into direct_message, one is to
the message that is being replied and the other one is to the message that is replying.

As it wasn't checking which of them was the replied one, it was editing the
reply message by mistake.