webhook message: Fix mixed messages with users having same name
When two users have the same username or nickname in a guild, webhook messages
get mixed together because nick#0000 matches as intended in discord, but
those webhook messages should be separated because they're two different users.

This was done to avoid misunderstanding and impersonating to people while talking in
these linked channels.

Now the following format will be applied "username#discriminator" but if user has
a nickname in guild, "nickname (username#discriminator)" will be applied instead.