Replies: Fix ping mention is superimposed when someone else talks
Sometimes when a user replies to a webhook message, the user mention
mentions the webhook, so while someone else talks in a channel, these
might look like they change to another user mention.

Therefore it'll be using the same way that webhook message usernames are
applied, not mention but text instead.